Fix Dialog so it doesn't exceed parent dimensions
Previously, when the height of the Dialog exceeded the parent,
the overflow would get cut off rather than allowing the user to
scroll through the entire contents of the Dialog.

There were several reasons why this was happening:
1) The height was being set to 100% (aka: pulled from the parent iframe) and
the height of the parent was already incorrect (aka: exceeded it's
parent).
2) Scroll was not enabled/supported.
3) The Dialog used position fixed in several places which made auto
positioning difficult.